Refactored actionFrame to manager
Refactored actionFrame to ActionManager class, which creates and manages tabs with their own gui setup.

Also refactored construction of resources to be inside ResourceManager. Kind of hate this but ActionManager needs access to an instance of ResourceManager to bind actions to buttons, which means that I either instantiate the GUI elements outside the managers and make them functional inside (which sounds awful and confusing), or just build the GUI elements and their containers inside the managers that need them (messy but maximizes independence).